create getthreads_run.js
this counts amount of threads available and runs a n1/n2 file that is specified at max capacity. Saves a lot of time and should be run from the home comp(theoretically you can run this for any comp but I suggest the home comp since it would have the required file to copy).